updating some of the fields to be decimals from int

+++ PecTec.Riot.LoL/Models/Champions/Info.cs
@@ -2,9 +2,9 @@
 {
     public class Info
     {
-        public int Attack { get; set; }
-        public int Defense { get; set; }
-        public int Magic { get; set; }
-        public int Difficulty { get; set; }
+        public decimal Attack { get; set; }
+        public decimal Defense { get; set; }
+        public decimal Magic { get; set; }
+        public decimal Difficulty { get; set; }
     }
 }
