+++ PecTec.Riot.LoL/Models/Champions/Specifics/Spell.cs
@@ -10,12 +10,12 @@ namespace PecTec.Riot.LoL.Models.Champions.Specifics
         public string Tooltip { get; set; }
         public Leveltip Leveltip { get; set; }
         public int Maxrank { get; set; }
-        public List<int> Cooldown { get; set; }
+        public List<decimal> Cooldown { get; set; }
         public string CooldownBurn { get; set; }
         public List<int> Cost { get; set; }
         public string CostBurn { get; set; }
         public Datavalues Datavalues { get; set; }
-        public List<List<int>> Effect { get; set; }
+        public List<List<decimal>> Effect { get; set; }
         public List<string> EffectBurn { get; set; }
         public List<object> Vars { get; set; }
         public string CostType { get; set; }
